Consolidate func code
