+++ src/compiler/typescript_to_rust/azle_generate/src/candid_type/azle_type_ref/to_data_type.rs
@@ -1,17 +1,10 @@
 use cdk_framework::act::node::{
-    candid::{Func, Opt, Primitive, TypeRef, Variant},
-    node_parts::mode::Mode,
+    candid::{Opt, Primitive, TypeRef},
     CandidType,
 };
 
 use super::AzleTypeRef;
-use crate::{
-    candid_type::func,
-    ts_ast::{
-        azle_type::AzleType, traits::GetTsType, AzleFnOrConstructorType,
-        FunctionAndMethodTypeHelperMethods, GetName,
-    },
-};
+use crate::ts_ast::traits::GetName;
 
 impl AzleTypeRef<'_> {
     pub fn to_data_type(&self) -> CandidType {
@@ -43,58 +36,6 @@ impl AzleTypeRef<'_> {
 }
 
 impl AzleTypeRef<'_> {
-    pub fn to_func(&self, name: Option<String>) -> Func {
-        let request_type_type_ref = match self.get_enclosed_azle_type() {
-            AzleType::AzleTypeRef(azle_type_ref) => azle_type_ref,
-            _ => panic!("{}", self.wrong_enclosed_type_error()),
-        };
-
-        let mode = match request_type_type_ref.get_name() {
-            "Query" => Mode::Query,
-            "Update" => Mode::Update,
-            "Oneway" => Mode::Oneway,
-            _ => panic!("{}", self.wrong_enclosed_type_error()),
-        };
-
-        let azle_fn_type = match request_type_type_ref.get_enclosed_azle_type() {
-            AzleType::AzleFnOrConstructorType(fn_or_const) => match fn_or_const {
-                AzleFnOrConstructorType::AzleFnType(ts_fn_type) => ts_fn_type,
-            },
-            _ => panic!("{}", self.wrong_enclosed_type_error()),
-        };
-
-        let params: Vec<CandidType> = azle_fn_type
-            .get_param_types()
-            .iter()
-            .map(|param| {
-                let azle_param = AzleType::from_ts_type(param.clone(), self.source_map);
-                azle_param.to_data_type()
-            })
-            .collect();
-
-        let return_type = AzleType::from_ts_type(
-            azle_fn_type.get_ts_type_ann().get_ts_type(),
-            self.source_map,
-        )
-        .to_data_type();
-
-        let to_vm_value = |name: String| func::generate_into_vm_value_impl(name);
-        let list_to_vm_value = |name: String| func::generate_list_into_vm_value_impl(name);
-        let from_vm_value = |name: String| func::generate_from_vm_value_impl(name);
-        let list_from_vm_value = |name: String| func::generate_list_from_vm_value_impl(name);
-
-        Func::new(
-            name,
-            params,
-            return_type,
-            mode,
-            to_vm_value,
-            list_to_vm_value,
-            from_vm_value,
-            list_from_vm_value,
-        )
-    }
-
     fn to_option(&self) -> Opt {
         let enclosed_act_data_type = self.get_enclosed_azle_type().to_data_type();
         Opt {
