+++ src/compiler/typescript_to_rust/azle_generate/src/ts_ast/azle_type/azle_type_lit/mod.rs
@@ -1,5 +1,3 @@
-use cdk_framework::act::node::candid::{variant, Variant};
-
 use swc_common::SourceMap;
 use swc_ecma_ast::TsTypeLit;
